Removed unnecessary things
ScopedPointer is removed since it could not handle containers. Use boost::shared_ptr instead, if required. Also removed AIModuleLoader since it wouldn't compile and isn't really necessary (we use a DLL anyway).

+++ FnulAIModule/Source/Commander/Squad.cpp
@@ -192,7 +192,7 @@ void Squad::computeActions()
 	checkAttack();
 
 	// TODO: Remove this
-	for (int i = 0; i < agents.size(); ++i)
+	for (size_t i = 0; i < agents.size(); ++i)
 	{
 		if (agents[i]->isUnit())
 		{
